Fix problem 59 chapter 1 case c solution
Previous provided solution for case c overcounts.
For example on 2 childs and 4 bars. Let a,b,c,d be bars.
If we use provided solution of multiplying amount of ways to divide
interchangeable bars between children and multiplying by 15! possible
permutations
(a, b), (c, d) and (b, a), (c, d)
divisions of bars will be counted as different, despite permutation of
a and b don't change which bars go to which child.

New solution just counts using ordered sampling with replacment.
Because each of 15 bars can be given to any of 10 children
we have 10^15 combinations.