Update 40.tex
The current justification for part a is incorrect - in the example given, P(X=Y) only equals 1 conditional on the information that Tom answers the question first. If the unconditional probability that person 1 answers Harare is 3/4, which is less than 1, there must be some situations where they do not answer Harare, which means there must be some situations where Tom does not answer the question first. 

The explanation I have provided uses a proof by contradiction to show that even a single point of divergence in the PMFs of X and Y must result in a nonzero chance that X and Y are not equal.

The explanation for part b is also incorrect - in the extreme case of random variables that always take some single value with probability 1, it is possible for two such variables to be technically independent and have probability 1 of being equal.