Update 10.tex
Current given reasoning is incorrect and asserts that, if X is the number of rounds, E(X^2) = (EX)^2, obtaining an answer of 4 for part 2, which is incorrect (verifiable by adding up X*P(X=K) for k from 1 to 5). A more detailed solution for part 2 using the derivation trick outlined in the textbook is instead provided.